Evict DecodingError and StopDecodingError
New home! lib/decode/errors!
