+++ packages/codec/lib/allocate/storage.ts
@@ -5,7 +5,8 @@ import { StoragePointer } from "../types/pointer";
 import { StorageAllocations, StorageAllocation, StorageMemberAllocation } from "../types/allocation";
 import { StorageLength, Range } from "../types/storage";
 import { isWordsLength } from "../utils/storage";
-import { UnknownUserDefinedTypeError, DecodingError } from "../types/errors";
+import { UnknownUserDefinedTypeError } from "../types/errors";
+import { DecodingError } from "../decode/errors";
 import { AstDefinition, AstReferences } from "../types/ast";
 import * as CodecUtils from "../utils";
 import { TypeUtils } from "../utils";
