+++ packages/codec/lib/index.ts
@@ -12,8 +12,9 @@ export { Utils };
 export * from "./interface";
 
 //now... various low-level stuff we want to export!
-//the actual decoding functions
+//the actual decoding functions and related errors
 export { decodeVariable, decodeEvent, decodeCalldata } from "./core/decoding";
+export { DecodingError, StopDecodingError } from "./decode/errors";
 
 //the debugger needs to get its allocations, and deal with storage sizes
 export { UnknownBaseContractIdError, getStorageAllocations, storageSize } from "./allocate/storage";
