+++ packages/codec/lib/read/stack.ts
@@ -2,7 +2,7 @@ import debugModule from "debug";
 const debug = debugModule("codec:read:stack");
 
 import * as CodecUtils from "../utils";
-import { DecodingError } from "../types/errors";
+import { DecodingError } from "../decode/errors";
 
 export function readStack(stack: Uint8Array[], from: number, to: number): Uint8Array {
   if(from < 0 || to >= stack.length) {
